Renamed some methods within the FOSS layer's toolchain recipes
There's a problem where sometimes Bitbake has parsed the old
openembedded/ layer external-toolchain-csl.bb recipe which
has the exact same methods as the one in the .inc in the
FOSS layer and it complains with an ERROR but doesn't stop
the build.  Added "caf_" in front of the methods to work
around the issue.

Change-Id: Ifae556c24abd00e2bef52686da44cec9cc119b20