Fix ramdisk generation for mdm
busybox/files/geninit file contained env variables
that would not expand when running gen_init_cpio.
When gen_init_cpio runs it exits with segfault when
env variables are in geninit file. This fix eliminates
the location information in env variables of files for
initial rootfs from geninit and moves that location
information to the build system instead.

Ramdisk requires busybox to be built as static. Configuration
files for ramdisk are now copied into a temp directory and
gen_init_cpio is executed from that directory.

Change-Id: Iec7c7dd243a3da10002df6b8926922ded3561a5d