Fix dependency issues for kernel and kernel modules.
Changes to the kernel will be reflected in $KERNEL_VERSION.
Add $KERNEL_VERSION to the PR for kernel module recipes.
This tells bitbake to rebuild modules when $KERNEL_VERSION changes
due to changes in the kernel.

Kernel module ipks were being created with a dependency
on a kernel ipk with the version included in the name.
This change renames the kernel ipk to include this version.

Change-Id: Ie63898fbd2c8c956848e47c539c2e0105c658d30