meta-qti-bsp: Adds module signing and aarch64 support to wlan
Adds kernel module signing and support for 64 bit targets to the
qcacld-ll recipe file. Previously the signing was not performed
automatically and the module wouldn't load if the kernel signing option
was enabled. Also, on 64 bit targets the binaries and configuration
files were placed into /lib64 instead of the /lib folder where they're
expected.

Change-Id: Ie0d8c3d7493b13f4968a135ef0a32ff93a13ee97