meta-qti-bsp: Mount /misc partition as emmc for apq8009
As on today, the recovery table is populating wrong entries
for /misc partition. Recovery binary needs to access the
/misc partition in write access mode to write the BCB in
Power failure / Battery Drain use cases. But due to wrong
entry of /misc, Recovery is not able to write to /misc.
This is leading the Power Failure Use case failure.
This has been fixed by updating the recovery volume of
/misc as emmc.

Change-Id: I2812a81b8370a51e42bc2667601f05a62f5f721a